Move tests to test_quoting.py
